Modifies Formula details --json for indented output. Fixes #26
